Add pure ruby punycode implementation to comply with GPL use
libidn2 is GPL. Apache 2.0 projects can't exclusively rely on GPL
dependencies, unless the license changes in those circumstances. So in
order to comply with the requirement, a pure ruby punycode
immplementation (IDNA 2003) is added in case the dependencies aren't
installed, to guarantee reasonable level of quality in the absence of
the other deps.